21 Mar 2023

+++ linearAlgebra/linear_recurrence_with_constant_coefficients.cpp
@@ -1,4 +1,6 @@
-int linear_recurrence_with_constant_coefficients(std::vector<int> A, std::vector<int> f, int n) {
+#include <vector>
+long long
+linear_recurrence_with_constant_coefficients(std::vector<int> A, std::vector<int> f, int n) {
     size_t k = A.size();
     Polynomial a(k + 1), b(k), c(k);
     for (int i = 0; i < k; ++i) {
@@ -18,4 +20,3 @@ int linear_recurrence_with_constant_coefficients(std::vector<int> A, std::vector
     }
     return res;
 }
-
