+++ string/minimal_representation.cpp
@@ -1,7 +1,7 @@
 #include <algorithm>
 #include <cstring>
 int minimal_representation(char *S) {
-    int n = strlen(S);
+    int n = (int) strlen(S);
     int i = 0, j = 1, k = 0;
     while (i < n && j < n && k < n) {
         if (S[(i + k) % n] == S[(j + k) % n]) {
